# Add support for res-block-fusing opt for 512 filter networks on Ampere