--- a/src/neural/cuda/common_kernels.cu
+++ b/src/neural/cuda/common_kernels.cu
@@ -546,6 +546,27 @@ void PolicyMap(int N, T* output, const T* input, const short* indices,
   ReportCUDAErrors(cudaGetLastError());
 }
 
+template <typename T = float, bool use_se, bool relu, bool use_bias, bool use_skip>
+void OutputInputTransform(int N, int C, int se_K, float* output,
+                          const float* input, const float* skip,
+                          const float* bias, const float* w1, const float* b1,
+                          const float* w2, const float* b2,
+                          cudaStream_t stream) {
+  // Each thread processes entire chess board
+  if (C > kMaxSupportedChannelsForResBlockFusing) {
+      throw Exception(
+          "res block fusing opt not supported for the given data type and no "
+          "of filters\n");
+  } else {
+    OutputTransform_SE_relu_InputTransform_kernel<float, use_se, relu, use_bias,
+                                                  use_skip>
+        <<<N, C, 0, stream>>>(N, C, se_K, output, input, (float*)skip, bias, w1,
+                              b1,
+                              w2, b2);
+  }
+  ReportCUDAErrors(cudaGetLastError());
+}
+
 // Template instantiation.
 template void copyTypeConverted<half, float>(half* op, float* ip, int N,
                                              cudaStream_t stream);