--- a/src/neural/cuda/common_kernels.cu
+++ b/src/neural/cuda/common_kernels.cu
@@ -546,7 +546,8 @@ void PolicyMap(int N, T* output, const T* input, const short* indices,
   ReportCUDAErrors(cudaGetLastError());
 }
 
-template <typename T = float, bool use_se, bool relu, bool use_bias, bool use_skip>
+template <typename T = float, bool use_se, bool relu, bool use_bias,
+          bool use_skip>
 void OutputInputTransform(int N, int C, int se_K, T* output, const T* input,
                           const T* skip, const T* bias, const T* w1,
                           const T* b1, const T* w2, const T* b2,
@@ -560,8 +561,7 @@ void OutputInputTransform(int N, int C, int se_K, T* output, const T* input,
     OutputTransform_SE_relu_InputTransform_kernel<float, use_se, relu, use_bias,
                                                   use_skip>
         <<<N, C, 0, stream>>>(N, C, se_K, output, input, (float*)skip, bias, w1,
-                              b1,
-                              w2, b2);
+                              b1, w2, b2);
   }
   ReportCUDAErrors(cudaGetLastError());
 }
@@ -624,7 +624,8 @@ template void FilterTransform<float>(int N, int C, float* transformedFilter,
 
 template void InputTransform<float, true>(int N, int C,
                                           float* transformed_input,
-                                          const float* input, cudaStream_t stream);
+                                          const float* input,
+                                          cudaStream_t stream);
 
 template void InputTransform<float, false>(int N, int C,
                                            float* transformed_input,