--- a/src/neural/cuda/fp16_kernels.cu
+++ b/src/neural/cuda/fp16_kernels.cu
@@ -195,6 +195,272 @@ bool Se_Fp16_NHWC(int N, int C, int numFc1Out, half* output, const half* skip,
   return true;
 }
 
+// Get board for this thread from shared memory
+// We are just using shared memory to store local 
+// thread data in this kernel to help reduce some register pressure 
+// and spills to local memory
+#define BOARD(y,x) shboard[(y) * 8 + (x)]
+
+// input is in transformed space (HWNC layout) --- output of GEMM
+// output is also in transformed space (HWNC layout) --- input to GEMM (for next layer)
+// 'C' threads per block
+// 'N' blocks
+// every thread generates an entire board/plane (8x8 elements)
+template <bool use_se, bool relu, bool use_bias, bool use_skip>
+__global__ __launch_bounds__(kMaxSupportedSeKForResBlockFusingFp16Ampere, 1)
+void OutputInputTransformKernel_fp16_shmem_board(int N, int C, int se_K, half* output, const half* input, 
+                                                 half* skip, const half* bias, const half* w1, const half* b1,
+                                                 const half* w2, const half* b2) {
+
+  int k = threadIdx.x;
+  int n = blockIdx.x;
+  
+  extern __shared__ half _sboard[];
+  half *shboard = &_sboard[k * 72];     // 72 instead of 64 to reduce shared memory bank conflicts
+
+  half b = bias[k];
+
+  #pragma unroll
+  for (int hStart = 0; hStart < 8; hStart += 4)
+    #pragma unroll
+    for (int wStart = 0; wStart < 8; wStart += 4) {
+      //  i) read to per thread registers (for doing output transform)
+      int shln = n * 4 + (hStart / 4) * 2 + (wStart / 4);
+      half outElTransformed[6][6];
+      #pragma unroll
+      for (int y = 0; y < 6; y++)
+        #pragma unroll
+        for (int x = 0; x < 6; x++)
+          outElTransformed[y][x] = input[TEMP_INDEX_HWNC(y, x, shln, k)];
+
+      // ii) transform it
+      half outEl[4][4];
+      OutputTransform4x4(&outEl[0][0], &outElTransformed[0][0]);
+
+      #pragma unroll
+      for (int y = 0; y < 4; y++)
+        copyAs<uint2>(&BOARD(hStart + y, wStart), &outEl[y][0]);
+    }
+
+  // Add bias, and compute the average for SE.
+  float S = 0;
+  float B = 0;
+
+  if (use_bias || use_se) {
+    #pragma unroll
+    for (int y = 0; y < 8; y++) {
+      half boardRow[8];
+      copyAs<uint4>(&boardRow, &BOARD(y, 0));
+      #pragma unroll
+      for (int x = 0; x < 8; x++) {
+        if (use_bias) boardRow[x] += b;
+        if (use_se) S += (float)boardRow[x];
+      }
+      if (use_bias) copyAs<uint4>(&BOARD(y, 0), &boardRow);
+    }
+  }
+
+  if (use_se) {
+    __shared__ float shared_data[kMaxSupportedSeKForResBlockFusingFp16Ampere];
+    float avg = S / 64;
+    shared_data[k] = avg;
+
+    int lane = k & 0x1F;
+    int warp = k >> 5;
+    __syncthreads();
+
+    // First fully-connected layer for SE
+
+    // As se_K << C, we want to loop over se_K instead of C
+    // even if it means taking the sum across threads
+
+    __shared__ float shared_sums[kMaxSupportedSeKForResBlockFusingFp16Ampere/32][kMaxSupportedSeKForResBlockFusing];  // per-warp sums
+
+    for (int i = 0; i < se_K; i++) {
+      float val = shared_data[k] * float(readw1(k, i));
+      val = warpReduce(val);
+      if (lane == 0)
+        shared_sums[warp][i] = val;
+    }
+    __syncthreads();
+    if (k < se_K) 
+    {
+      S = 0;
+      for (int i=0;i<C/32;i++)
+        S += shared_sums[i][k];
+
+      S += (float)b1[k];
+      if (S < 0) S = 0;  // relu
+      shared_data[k] = S;
+    }
+
+    __syncthreads();
+
+    // Second fully-connected layer for SE
+    S = 0;
+    for (int i = 0; i < se_K; i++) {
+      float val = shared_data[i];
+      S += val * float(readw2(i, k));
+      B += val * float(readw2(i, k + C));
+    }
+    S += (float)b2[k];
+    B += (float)b2[k + C];
+
+    // Sigmoid (only on the scale part).
+    S = 1.0f / (1.0f + exp(-S));
+  }
+
+
+  // Scale/bias, add skip connection, perform relu, and write to output.
+  if (use_se || use_skip || relu) {
+    for (int h = 0; h < 8; h++) {
+      half boardRow[8];
+      copyAs<uint4>(&boardRow[0], &BOARD(h, 0));
+
+      if (use_se) {
+        #pragma unroll
+        for (int w = 0; w < 8; w++) boardRow[w] = (half)(float(boardRow[w]) * S + B);
+      }
+
+      // residual add
+      if (use_skip) {
+        half skipInp[8];
+        copyAs<uint4>(&skipInp[0], &skip[INDEX_NHCW(n, k, h, 0)]);
+        #pragma unroll
+        for (int w = 0; w < 8; w++) boardRow[w] += skipInp[w];
+      }
+
+      // relu
+      if (relu) {
+        #pragma unroll
+        for (int w = 0; w < 8; w++)
+          if (boardRow[w] < (half)0) boardRow[w] = 0;
+      }
+
+      // write un-transformed output to 'skip' if required
+      if (use_skip)
+      {
+        copyAs<uint4>(&skip[INDEX_NHCW(n, k, h, 0)], &boardRow[0]);
+      }
+
+      copyAs<uint4>(&BOARD(h, 0), &boardRow);
+    }
+  }
+
+  // perform input transform    
+
+  int c = k;
+  // top-left
+  {
+    half inEl[6][6] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+
+    #pragma unroll
+    for (int i = 0; i < 5; i++)
+      #pragma unroll
+      for (int j = 0; j < 5; j++) inEl[i + 1][j + 1] = BOARD(i,j);
+
+    InputTransform4x4(&inEl[0][0], &inEl[0][0]);
+
+    #pragma unroll
+    for (int y = 0; y < 6; y++)
+      #pragma unroll
+      for (int x = 0; x < 6; x++)
+        output[TEMP_INDEX_HWNC(y, x, n * 4 + 0, c)] = inEl[y][x];
+  }
+
+  // top-right
+  {
+    half inEl[6][6] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+
+    #pragma unroll
+    for (int i = 0; i < 5; i++)
+      #pragma unroll
+      for (int j = 0; j < 5; j++) inEl[i + 1][j] = BOARD(i,j+3);
+
+    InputTransform4x4(&inEl[0][0], &inEl[0][0]);
+
+    #pragma unroll
+    for (int y = 0; y < 6; y++)
+      #pragma unroll
+      for (int x = 0; x < 6; x++)
+        output[TEMP_INDEX_HWNC(y, x, n * 4 + 1, c)] = inEl[y][x];
+  }
+
+  // bottom-left
+  {
+    half inEl[6][6] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+
+    #pragma unroll
+    for (int i = 0; i < 5; i++)
+      #pragma unroll
+      for (int j = 0; j < 5; j++) inEl[i][j + 1] = BOARD(i+3,j);
+
+    InputTransform4x4(&inEl[0][0], &inEl[0][0]);
+
+    #pragma unroll
+    for (int y = 0; y < 6; y++)
+      #pragma unroll
+      for (int x = 0; x < 6; x++)
+        output[TEMP_INDEX_HWNC(y, x, n * 4 + 2, c)] = inEl[y][x];
+  }
+
+  // bottom-right
+  {
+    half inEl[6][6] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+
+    #pragma unroll
+    for (int i = 0; i < 5; i++)
+      #pragma unroll
+      for (int j = 0; j < 5; j++) inEl[i][j] = BOARD(i+3,j+3);
+
+    InputTransform4x4(&inEl[0][0], &inEl[0][0]);
+
+    #pragma unroll
+    for (int y = 0; y < 6; y++)
+      #pragma unroll
+      for (int x = 0; x < 6; x++)
+        output[TEMP_INDEX_HWNC(y, x, n * 4 + 3, c)] = inEl[y][x];
+  } 
+}
+
+template <typename T = half, bool use_se, bool relu, bool use_bias, bool use_skip>
+void OutputInputTransform(int N, int C, int se_K, half* output,
+                          const half* input, const half* skip, const half* bias,
+                          const half* w1, const half* b1, const half* w2,
+                          const half* b2,
+                          cudaStream_t stream) {
+  // Each thread processes entire chess board
+  if (C > kMaxSupportedChannelsForResBlockFusing) {
+    // use special kernel with reduced register pressure - only works on Ampere,
+    // and only for fp16
+    if (C <= kMaxSupportedSeKForResBlockFusingFp16Ampere) {
+      cudaFuncSetAttribute(
+          OutputInputTransformKernel_fp16_shmem_board<use_se, relu, use_bias,
+                                                      use_skip>,
+          cudaFuncAttributeMaxDynamicSharedMemorySize, 72 * 1024);
+      OutputInputTransformKernel_fp16_shmem_board<use_se, relu, use_bias,
+                                                  use_skip>
+          <<<N, C, 72 * 1024, stream>>>(
+              N, C, se_K, (half*)output, (const half*)input, (half*)skip,
+              (half*)bias, (half*)w1, (half*)b1, (half*)w2, (half*)b2);
+    } else {
+      throw Exception(
+          "res block fusing opt not supported for the given data type and no "
+          "of filters\n");
+    }
+  } else {
+    OutputTransform_SE_relu_InputTransform_kernel<half, use_se, relu, use_bias,
+                                                  use_skip>
+        <<<N, C, 0, stream>>>(N, C, se_K, output, input, (half*)skip, bias, w1, b1,
+                              w2, b2);
+  }
+  ReportCUDAErrors(cudaGetLastError());
+}
+
 template void FilterTransform<half>(int N, int C, half* transformedFilter,
                                     const half* filter);