--- a/src/neural/cuda/kernels.h
+++ b/src/neural/cuda/kernels.h
@@ -36,7 +36,8 @@ void addVectors(T* c, T* a, T* b, int size, int asize, int bsize, bool relu,
 
 // Add bias to convolution's output.
 template <typename T>
-void addBias_NCHW(T* c, T* a, T* b, int N, int C, int H, int W, bool relu, cudaStream_t stream);
+void addBias_NCHW(T* c, T* a, T* b, int N, int C, int H, int W, bool relu,
+                  cudaStream_t stream);
 
 // Conversion from: fp32 -> fp16 datatype, and NCHW -> NHWC layout.
 // Cudnn kernels work best with NCHW layout for fp32, and with NHWC for fp16.
@@ -83,13 +84,13 @@ void PolicyMap(int N, T* output, const T* input, const short* indices,
                int inputSize, int usedSize, int outputSize,
                cudaStream_t stream);
 
-
 // Custom winograd helper functions
 template <typename T>
 void FilterTransform(int N, int C, T* transformedFilter, const T* filter);
 
 template <typename T, bool nhcw>
-void InputTransform(int N, int C, T* transformedInput, const T* input, cudaStream_t stream);
+void InputTransform(int N, int C, T* transformedInput, const T* input,
+                    cudaStream_t stream);
 
 template <typename T, bool use_se, bool relu, bool use_bias, bool use_skip,
           bool skipInput_nhcw, bool output_nhcw>
@@ -99,8 +100,9 @@ void OutputTransform(int N, int C, int se_K, T* output, const T* input,
 
 template <typename T, bool use_se, bool relu, bool use_bias, bool use_skip>
 void OutputInputTransform(int N, int C, int se_K, T* output, const T* input,
-                     const T* skip, const T* bias, const T* w1, const T* b1,
-                     const T* w2, const T* b2, cudaStream_t stream);
+                          const T* skip, const T* bias, const T* w1,
+                          const T* b1, const T* w2, const T* b2,
+                          cudaStream_t stream);
 
 }  // namespace cudnn_backend
 }  // namespace lczero