* Use a special kernel for handling > 384 filter res block fusion that allocates board[8][8] array in shared memory instead of registers.
 * This works only on Ampere GPUs because the kernel needs more than 64KB shared memory that is available on Turing.

Some performance nos:

```
backend bench with 512fx40b network:
=============

GA100:

baseline: Benchmark batch size 96 with inference average time 9.27067ms - throughput 10355.2 nps.
opt:      Benchmark batch size 96 with inference average time 6.319ms   - throughput 15192.3 nps.

RTX 3070:

baseline: Benchmark batch size 96 with inference average time 23.8637ms - throughput 4022.84 nps.
with opt: Benchmark batch size 96 with inference average time 17.8473ms - throughput 5378.97 nps.

RTX 3080:

baseline: Benchmark batch size 128 with inference average time 22.2339ms - throughput 5756.98 nps.
with opt: Benchmark batch size 128 with inference average time 17.4429ms - throughput 8011.02 nps.


for comparison, 384fx30b on GA100: 

baseline: Benchmark batch size 96 with inference average time 5.16775ms - throughput 18576.8 nps.
opt:      Benchmark batch size 96 with inference average time 3.36603ms - throughput 28520.2 nps.

```

Which means with the optimization 512fx40b network runs less than half as slow as 384fx30b network on Ampere.