@borg323 2021-11-28
I needed the following to build with clang:
```diff
diff --git a/src/neural/cuda/common_kernels.cu b/src/neural/cuda/common_kernels.cu
index 4dfe73c..a263ed2 100644
--- a/src/neural/cuda/common_kernels.cu
+++ b/src/neural/cuda/common_kernels.cu
@@ -547,10 +547,10 @@ void PolicyMap(int N, T* output, const T* input, const short* indices,
 }
 
 template <typename T = float, bool use_se, bool relu, bool use_bias, bool use_skip>
-void OutputInputTransform(int N, int C, int se_K, float* output,
-                          const float* input, const float* skip,
-                          const float* bias, const float* w1, const float* b1,
-                          const float* w2, const float* b2,
+void OutputInputTransform(int N, int C, int se_K, T* output,
+                          const T* input, const T* skip,
+                          const T* bias, const T* w1, const T* b1,
+                          const T* w2, const T* b2,
                           cudaStream_t stream) {
   // Each thread processes entire chess board
   if (C > kMaxSupportedChannelsForResBlockFusing) {
diff --git a/src/neural/cuda/fp16_kernels.cu b/src/neural/cuda/fp16_kernels.cu
index a3b99d8..ee1cede 100644
--- a/src/neural/cuda/fp16_kernels.cu
+++ b/src/neural/cuda/fp16_kernels.cu
@@ -428,10 +428,10 @@ void OutputInputTransformKernel_fp16_shmem_board(int N, int C, int se_K, half* o
 }
 
 template <typename T = half, bool use_se, bool relu, bool use_bias, bool use_skip>
-void OutputInputTransform(int N, int C, int se_K, half* output,
-                          const half* input, const half* skip, const half* bias,
-                          const half* w1, const half* b1, const half* w2,
-                          const half* b2,
+void OutputInputTransform(int N, int C, int se_K, T* output,
+                          const T* input, const T* skip, const T* bias,
+                          const T* w1, const T* b1, const T* w2,
+                          const T* b2,
                           cudaStream_t stream) {
   // Each thread processes entire chess board
   if (C > kMaxSupportedChannelsForResBlockFusing) {
```